feat: Walk Repository 구현 및 테스팅 완료

+++ src/main/java/com/kakaoseventeen/dogwalking/dog/DogJpaRepository.java
@@ -5,5 +5,5 @@ import java.util.List;
 
 public interface DogJpaRepository extends JpaRepository<Dog, Integer> {
 
-    List<Dog> findDogsByMemberId(int memberId);
+    List<Dog> findDogsByMemberId(long memberId);
 }
